Feature Implementation and Bug Fixes
1. .bashrc wasn't reading environmental variables if the changed command
was in the file. Needs to be commented out. (Ubuntu 16.04).

2. Added ability to set the port type. If you're trying to change a
whole domain A like domain.com the @ symbol in vesta gets used a lot.
Without looking for the A record, it picks up NS records. Made the
ability to set which record to change.

3. Updated the remote Dynamic IP grabber code. Was an issue on some
machines where the IP came back with the IP of the Vesta Server.